[NSObject] Unify some code between iOS and macOS.
* Add these methods to shared code so they're available on all platforms
  (they're already available on mobile platforms), since there's no reason to
  exclude them on macOS:
    * NSObject.Init
    * NSObject.Alloc
    * NSObject.InvokeInBackground
* Remove unused usings.
* Move identical code in platform-specific files to shared code.
